requirements.txt
# Core Web Frameworks
streamlit>=1.28.0

langchain>=1.2.10
langgraph>=1.0.8
langchain-google-genai>=4.2.0

# HTTP Clients
requests>=2.31.0

# Vector Database & Embeddings
chromadb>=0.4.15
sentence-transformers>=2.2.2

# Data Validation
pydantic>=2.5.0

# PyTorch CPU-only (for Streamlit Cloud deployment)
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
torchvision>=0.15.0

# ML & Image Processing
timm>=0.9.0
pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
opencv-python-headless>=4.8.0
grad-cam>=1.4.8
pandas>=2.0.0
matplotlib>=3.7.0
reportlab>=4.4.10

# MedGemma / Critic Agent
transformers>=4.38.0
accelerate>=0.27.0
huggingface_hub>=0.20.0
gradio_client>=0.8.0

ultralytics>=8.0.0
ipykernel>=6.0.0
